[refactor] Apply command pattern
- create shell_command.py
- refactor source code architecture
